refactor(headscale): removal of headscale
Currently headscale is not used beeing deployed & used with Greenhouse.
To avoid unnecessary refactorings and time spend to validate Dependabot PRs
we decided to not touch the code for now.
This PR removes all usage of Headscale, so that we can revert when we have
a concrete usage for this in the future.