feat(plugin): add ReleaseName field and defaulting logic
Helm limits the length of the release name to 53 chars
Currently the Plugin name is used as the releaseName.
With PluginPresets this name consists of PluginPreset name
+ Cluster name, which can get longer.

This PR changes the behavior for new Plugins to have either
the releaseName from the PluginSpec or use the PluginDefinitions
HelmChart name.
The name is limited to 53 characters in both cases.

- for existing Plugins the Name of the plugin will be used
- for newly created Plugins the ReleaseName is used if set, else the name of the HelmChart is used
- releaseName is optional until set