fix(teamroles): add fieldindex in webhook setup
After the Split of Controller and Webhook deployment the deletion
of TeamRoles was failing due to a missing FieldIndex

This add the missing field indexing to the teamrole webhook setup.
To avoid conflicts with already existing indices the IgnoreIndexerConflict(..) was added.
The code does not expose a typed error, so in this case the check
relies on the error's prefix