feat(plugins): add ReleaseName field and defaulting logic (#1121)
* feat(plugin): add ReleaseName field and defaulting logic

Helm limits the length of the release name to 53 chars
Currently the Plugin name is used as the releaseName.
With PluginPresets this name consists of PluginPreset name
+ Cluster name, which can get longer.

This PR changes the behavior for new Plugins to have either
the releaseName from the PluginSpec or use the PluginDefinitions
HelmChart name.
The name is limited to 53 characters in both cases.

- for existing Plugins the Name of the plugin will be used
- for newly created Plugins the ReleaseName is used if set, else the name of the HelmChart is used
- releaseName is optional until set

* Automatic generation of CRD API Docs

* set releaseName for service-proxy

* regenerate crds

* ensure releaseName is compliant with helm

* ensure helm release used consistently

* Automatic generation of CRD API Docs

* regenerate api-docs

* Automatic generation of CRD API Docs

* fix test method

* remove unused variable

* fix(pluginpresets): ensure update of existing plugins

when the PluginPreset defines a releaseName it must be ensured
that the releaseName for pre-existing plugins is not changed

---------

Co-authored-by: Cloud Operator <169066274+cloud-operator@users.noreply.github.com>