feat(core): add annotation to trigger reconcile
The new annotation 'greenhouse.sap/operation' should be used
to trigger certain operations on the resources.

First supported operation is `reconcile` which is there to trigger
the reconciliation of an object. The handling is added to the central
lifecyle package, where the annotation is removed before continuing
with the actual reconciliation loop of the respective resource