feat(pluginpresets): add deletionPolicy support
Introduces `.spec.deletionPolicy` on the PluginPreset.
The default policy is `Delete` which results in the Cleanup of all Plugins
owned by the PluginPreset.
The policy `Orphan` can be used if it is desired to keep the Plugins. The
label key `greenhouse.sap/pluginpreset` is kept on the Plugin resource, so
that creating the same Preset again will ensure the Plugins are managed again.
Only the OwnerReference is removed during the deletion with Orphan