feat(core): support reconcile annotations
the 'greenhouse.sap/reconcile' annotation triggers reconciliation
also on resources managed by a Greenhouse controller.
For Plugins this will set the appropriate reconciledAt annoations
on the flux.HelmRelease CR.
The current value of this annotation is saved to the status, to
Once the annotation is removed the status field is cleared.

Updating of the status field has been added to the lifecycle pkg.
In order to support this new field the Greenhouse resource must
satisfy the LastReconcileSetter interface.
Custom logic may be integrated into the existing EnsureCreated
method