docs(core): rework of documentation
This PR aims to rework the documetation structure.
getting-started/core-concepts should not contain
technical details but give a brief overview.
How to configure the API resources should be shown
under reference/components. This is a more detailed
documentation and may contain some additional info
on how to use the API. Next to the generated api docs.
user-guides should only contain examples/tutorials
that guide the user to complete a specific task.

to make the navigation and finding related resources
easier the refactored docs pages will now have a
'Next steps' section at the bottom of the page