docs(core): rework of documentation (#1558)
* docs(core): rework of documentation

This PR aims to rework the documetation structure.
getting-started/core-concepts should not contain
technical details but give a brief overview.
How to configure the API resources should be shown
under reference/components. This is a more detailed
documentation and may contain some additional info
on how to use the API. Next to the generated api docs.
user-guides should only contain examples/tutorials
that guide the user to complete a specific task.

to make the navigation and finding related resources
easier the refactored docs pages will now have a
'Next steps' section at the bottom of the page

* Update docs/getting-started/core-concepts/organizations.md

Co-authored-by: Krzysztof Zagorski <k.zaggy@gmail.com>

* fix(docs): remove .md suffix from links

* fix: links & code review

* fix: more broken links

* fix: more broken links

* fix: broken links

On-behalf-of: @SAP krzysztof.zagorski@sap.com

* point to new location for seeded rbac

* add reference to vault/openbao reference epic

* fix: plugin example option valueFrom secret

---------

Co-authored-by: Krzysztof Zagorski <k.zaggy@gmail.com>