fix(plugins): ensure SecretRefs are available for templating (#1680)
* fix(plugins): ensure SecretRefs are available for templating

Templating of the Plugins HelmChart could fail if a required SecretRef was not available.
This change ensures that secretRefs are not filtered from the optionValues during evaluation of expression & integrated values. Also it ensures they are not removed when calculating the flux helmrelease values

* refactor(plugins): process PluginOptionValues individually

This refactors the resolution of ValueFromRefs and Expressions
Instead of individual processing of optionValues incl filtering and merging of results.
this is now processing each value individually handling CEL expressions and
tracking of integrated values